fix: premature resolution of wait conditions when using scopedBy
When using a function style selector wait conditions fail early if the
scoped selector does not yield any results. Instead we should continue
to wait/check until the wait condition times out.

Additionally in order to retain custom friendly error messages that
include the selector that an exception occurred on we fallback to the
original wait behavior which includes the custom errors.