fix: kill all remaining webdriver browsers
An orphaned webdriver browser would only get its ppid set to 1 in mac.
In linux, it would get set to systemd, which is not 1.
We can't rely on that number, so instead we just kill any left over browsers spawned by webdriver,
even if another theoretical webdriver instance is using them.