fix: call `waitForPromisesToFlushBetweenTests` in `setUpWebDriverBeforeEach`
I saw in some logs the "Writing failure artifact to ..." line in the middle of a getPageSource output. I think it was the failure output of a `beforeEach` running at the same time as the failure output of an `it`. This is because failureArtifacts calls in `beforeEach` don't get awaited until `setUpWebDriverAfterEach`. If we can await the code before then next `it` happens, we may be able to prevent this.