feat: store original test error in failure artifact
When using retries, if the test retries to success, it is hard to find the original test error. You have to view the logs with `DEBUG=@faltest/mocha` and look for `@faltest/mocha Retrying failed test "my test": my error`. This makes it a lot easier to diagnose a flakey error.